backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]

# AI/ML (Anthropic only — journal companion, patterns, synthesis)
anthropic>=0.40.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
slowapi==0.1.9

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0

# Structured Logging (optional)
structlog>=23.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
